chore: move _internal.axis_transform to axis.transform
